requirements.txt
altair==4.1.0
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.3.4
astor==0.8.1
astroid==2.4.2
async-generator==1.10
attrs==20.3.0
autopep8==1.5.6
backcall==0.2.0
base58==2.1.0
beautifulsoup4==4.9.3
bleach==3.2.1
blinker==1.4
bs4==0.0.1
cachetools==4.2.1
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.4
confuse==1.4.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
Django==3.1.8
django-cors-headers==3.7.0
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-jsonapi==4.1.0
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
docopt==0.6.2
dominate==2.6.0
driver==0.1.0
entrypoints==0.3
et-xmlfile==1.0.1
filelock==3.0.12
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gitdb==4.0.7
GitPython==3.1.14
htmlmin==0.1.12
idna==2.10
ImageHash==4.2.0
inflection==0.5.1
install==1.3.4
ipykernel==5.3.4
ipympl==0.6.3
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.6.4
itsdangerous==1.1.0
itypes==1.2.0
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.2
joblib==1.0.1
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab==2.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
lxml==4.6.1
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.6.1
missingno==0.4.2
mistune==0.8.4
mord==0.6
mysql==0.0.2
mysqlclient==2.0.1
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
networkx==2.5.1
notebook==6.1.5
notebook-as-pdf==0.4.0
numpy==1.19.3
oauthlib==3.1.0
openai==0.2.6
opencv-python==4.4.0.46
openpyxl==3.0.5
packaging==20.4
pandas==1.2.0
pandas-profiling==2.11.0
pandocfilters==1.4.3
parso==0.7.1
patsy==0.5.1
pexpect==4.8.0
phik==0.11.2
pickleshare==0.7.5
Pillow==8.0.1
pipreqs==0.4.10
prettytable==2.1.0
prometheus-client==0.9.0
prompt-toolkit==3.0.8
protobuf==3.15.8
ptyprocess==0.7.0
pyarrow==3.0.0
pycodestyle==2.7.0
pycparser==2.20
pydeck==0.6.2
pyee==8.1.0
Pygments==2.7.2
PyJWT==2.0.1
pylint==2.6.0
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.6
pyparsing==2.4.7
PyPDF2==1.26.0
pyppeteer==0.2.5
pyrsistent==0.17.3
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.5
PyWavelets==1.1.1
pywin32==300
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==20.0.0
qtconsole==4.7.7
QtPy==1.9.0
requests==2.25.0
requests-oauthlib==1.3.0
scikit-learn==0.24.1
scipy==1.6.0
seaborn==0.11.1
selenium==3.141.0
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
smmap==4.0.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.0.1
SQLAlchemy==1.3.20
sqlparse==0.4.1
statsmodels==0.12.2
streamlit==0.80.0
Strip==0.0.1.dev14
tabulate==0.8.9
tangled-up-in-unicode==0.0.7
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.59.0
traitlets==5.0.5
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.2
validators==0.18.2
virtualenv==20.4.4
visions==0.6.0
visitor==0.1.3
watchdog==2.0.2
wcwidth==0.2.5
webencodings==0.5.1
websockets==8.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
WTForms==2.3.3
yarg==0.1.9
